Add script for near-print board
Run this on your board connected to your printer. This one checks if the server has received sth that has to be printed and uploads it, sends to the printer and deletes the downloaded file from the server, afterwards just keeps monitoring if the server got something to print.

Sending for a print will be added later. The current version is just monitoring and downloading files to a directory